+++ core/src/io/anuke/mindustry/editor/MapEditor.java
@@ -48,7 +48,7 @@ public class MapEditor{
         if(clear){
             for(int x = 0; x < map.width(); x++){
                 for(int y = 0; y < map.height(); y++){
-                    map.write(x, y, DataPosition.floor, (byte) Blocks.stone.id);
+                    map.write(x, y, DataPosition.floor, Blocks.stone.id);
                 }
             }
         }
@@ -277,7 +277,7 @@ public class MapEditor{
         map = new MapTileData(width, height);
         for(int x = 0; x < map.width(); x++){
             for(int y = 0; y < map.height(); y++){
-                map.write(x, y, DataPosition.floor, (byte) Blocks.stone.id);
+                map.write(x, y, DataPosition.floor, Blocks.stone.id);
             }
         }
         renderer.resize(width, height);
